Docker

+++ docker_start.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
 ls -la /config
-cp -R -u -p /src/app/config/global_dist.js /config/config.js
+
+if [ ! -f /config/config.js ]; then
+    cp -R -u -p /src/app/config/global_dist.js /config/config.js
+fi
 
 ln -s /config/config.js /src/app/config/global.js
